+++ examples/plugin-site/gatsby-config.js
@@ -1,8 +1,8 @@
 module.exports = {
   siteMetadata: {
-    title: 'gatsby-plugin-notionso example site',
+    title: 'gatsby-source-notionso example site',
     description:
-      'series of article to explain how to use gatsby-plugin-notionso',
+      'series of article to explain how to use the gatsby-source-notionso',
     basePath: '/',
   },
   plugins: [
@@ -10,8 +10,7 @@ module.exports = {
       resolve: 'gatsby-source-notionso',
       options: {
         name: 'Blog',
-        rootPageUrl:
-          'https://www.notion.so/gatsby-source-notionso-80562eadd8334396a905c372791cbf1b',
+        rootPageUrl: process.env.GATSBY_NOTIONSO_ROOT_PAGE_URL,
         debug: false,
       },
     },