+++ CHANGELOG.md
@@ -3,11 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## unpublished
+## 1.0.0
 ### Added
 - add an example in `examples` directory
-
-
+- publish site at http://www.gatsbyplugins.com
+- add netlify automation
 
 ## 0.0.6 - 2020-01-11
 ### Changed